Completed shader parser.
Files can be loaded and formatted with the parser. This separates files
into different shader types and also adds include functionality,
enabling multiple files for a single shader.

+++ GameEngine/GameEngine/src/Application.cpp
@@ -66,7 +66,7 @@ namespace engine
 
   void Application::Frame()
   {
-    s_context->totalDeltaTime += static_cast<float>(s_context->frameTime.getSeconds());
+    s_context->totalDeltaTime += s_context->frameTime.getSeconds();
     s_context->frameTime.Reset();
 
     if (s_context->nextScene)
